steering calibration when output from emergency

+++ arduino_codes/clear_ros_node/headers/encoder_hardware_interface.h
@@ -34,7 +34,7 @@ public:
 
     byte encoderReset(int encNum);
 
-    byte encoderWrite(byte encNum, long count);
+    byte encoderWrite(int encNum, long count);
 
 };
